refactor: reorganize url template utilities and improve memoization
