+++ src/url-template/url-template.ts
@@ -139,7 +139,7 @@ function getValues(
   return result;
 }
 
-function parseTemplate(template: string): Template {
+export function parseTemplate(template: string): Template {
   const operators = ["+", "#", ".", "/", ";", "?", "&"];
 
   return {
@@ -189,30 +189,3 @@ function parseTemplate(template: string): Template {
     },
   };
 }
-
-function memoize<K, V>(fn: (_key: K) => V): (_key: K) => V {
-  const cache = new Map<K, V>();
-
-  return (key: K) => {
-    if (cache.has(key)) {
-      return cache.get(key) as V;
-    }
-
-    const val = fn(key);
-    cache.set(key, val);
-
-    return val;
-  };
-}
-
-interface Expander {
-  // eslint-disable-next-line no-unused-vars -- This is actually used
-  expand(parameters: unknown): string;
-}
-
-// eslint-disable-next-line no-unused-vars -- This is actually used
-type Parse = (pathTemplate: string) => Expander;
-
-export const parseUrlTemplate: Parse = memoize(
-  (pathTemplate: string) => parseTemplate(pathTemplate)
-);
